+++ sympy/physics/quantum/operatorordering.py
@@ -1,6 +1,4 @@
-"""Bosonic and fermionic quantum operators.
-
-"""
+"""Functions for reordering operator expressions."""
 
 from warnings import warn
 
@@ -9,211 +7,12 @@ from sympy import Add, Mul, Pow, Integer, exp, sqrt, conjugate
 from sympy.physics.quantum import Operator, Commutator, AntiCommutator, Dagger
 from sympy.physics.quantum import HilbertSpace, FockSpace, Ket, Bra
 from sympy.functions.special.tensor_functions import KroneckerDelta
-
+from sympy.physics.quantum import BosonOperator, FermionOperator
 
 __all__ = [
-    'BosonOperator',
-    'BosonFockKet',
-    'BosonFockBra',
-    'BosonCoherentKet',
-    'BosonCoherentBra',
-    'FermionOperator',
-    'FermionFockKet',
-    'FermionFockBra',
     'normal_order',
     'normal_ordered_form'
 ]
-
-
-class BosonOperator(Operator):
-    """A bosonic operator that satisfies [a, Dagger(a)] == 1.
-
-    Parameters
-    ==========
-
-    name : str
-        A string that labels the bosonic mode.
-
-    annihilation : bool
-        A bool that indicates if the bosonic operator is an annihilation (True,
-        default value) or creation operator (False)
-
-    Examples
-    ========
-
-    >>> from sympy.physics.quantum import Dagger, Commutator, BosonOperator
-    >>> a = BosonOperator("a")
-    >>> Commutator(a, Dagger(a)).doit()
-    1
-    """
-
-    @property
-    def name(self):
-        return self.args[0]
-
-    @property
-    def is_annihilation(self):
-        return bool(self.args[1])
-
-    @classmethod
-    def default_args(self):
-        return ("a", True)
-
-    def __new__(cls, *args, **hints):
-        if not len(args) in [1, 2]:
-            raise ValueError('1 or 2 parameters expected, got %s' % args)
-
-        if len(args) == 1:
-            args = (args[0], Integer(1))
-
-        if len(args) == 2:
-            args = (args[0], Integer(args[1]))
-
-        return Operator.__new__(cls, *args)
-
-    def _eval_commutator_BosonOperator(self, other, **hints):
-        if self.name == other.name:
-            # [a^\dagger, a] = -1
-            if not self.is_annihilation and other.is_annihilation:
-                return Integer(-1)
-
-        elif 'independent' in hints and hints['independent']:
-            # [a, b] = 0
-            return Integer(0)
-
-        return None
-
-    def _eval_commutator_FermionOperator(self, other, **hints):
-        return Integer(0)
-
-    def _eval_anticommutator_BosonOperator(self, other, **hints):
-        if 'independent' in hints and hints['independent']:
-            # {a, b} = 0
-            return 2 * self * other
-
-        return None
-
-    def _eval_anticommutator_FermionOperator(self, other, **hints):
-        return 2 * self * other
-
-    def _eval_adjoint(self):
-        return BosonOperator(str(self.name), not self.is_annihilation)
-
-    def _print_contents_latex(self, printer, *args):
-        if self.is_annihilation:
-            return r'{%s}' % str(self.name)
-        else:
-            return r'{{%s}^\dag}' % str(self.name)
-
-    def _print_contents(self, printer, *args):
-        if self.is_annihilation:
-            return r'%s' % str(self.name)
-        else:
-            return r'Dagger(%s)' % str(self.name)
-
-    def _print_contents_pretty(self, printer, *args):
-        from sympy.printing.pretty.stringpict import prettyForm
-        pform = printer._print(self.args[0], *args)
-        if self.is_annihilation:
-            return pform
-        else:
-            return pform**prettyForm(u('\u2020'))
-
-
-class FermionOperator(Operator):
-    """A fermionic operator that satisfies {c, Dagger(c)} == 1.
-
-    Parameters
-    ==========
-
-    name : str
-        A string that labels the fermionic mode.
-
-    annihilation : bool
-        A bool that indicates if the fermionic operator is an annihilation
-        (True, default value) or creation operator (False)
-
-    Examples
-    ========
-
-    >>> from sympy.physics.quantum import Dagger, AntiCommutator
-    >>> from sympy.physics.quantum import FermionOperator
-    >>> c = FermionOperator("c")
-    >>> AntiCommutator(c, Dagger(c)).doit()
-    1
-    """
-    @property
-    def name(self):
-        return self.args[0]
-
-    @property
-    def is_annihilation(self):
-        return bool(self.args[1])
-
-    @classmethod
-    def default_args(self):
-        return ("c", True)
-
-    def __new__(cls, *args, **hints):
-        if not len(args) in [1, 2]:
-            raise ValueError('1 or 2 parameters expected, got %s' % args)
-
-        if len(args) == 1:
-            args = (args[0], Integer(1))
-
-        if len(args) == 2:
-            args = (args[0], Integer(args[1]))
-
-        return Operator.__new__(cls, *args)
-
-    def _eval_commutator_FermionOperator(self, other, **hints):
-        if 'independent' in hints and hints['independent']:
-            # [c, d] = 0
-            return Integer(0)
-
-        return None
-
-    def _eval_anticommutator_FermionOperator(self, other, **hints):
-        if self.name == other.name:
-            # {a^\dagger, a} = 1
-            if not self.is_annihilation and other.is_annihilation:
-                return Integer(1)
-
-        elif 'independent' in hints and hints['independent']:
-            # {c, d} = 2 * c * d, because [c, d] = 0 for independent operators
-            return 2 * self * other
-
-        return None
-
-    def _eval_anticommutator_BosonOperator(self, other, **hints):
-        # because fermions and bosons commute
-        return 2 * self * other
-
-    def _eval_commutator_BosonOperator(self, other, **hints):
-        return Integer(0)
-
-    def _eval_adjoint(self):
-        return FermionOperator(str(self.name), not self.is_annihilation)
-
-    def _print_contents_latex(self, printer, *args):
-        if self.is_annihilation:
-            return r'{%s}' % str(self.name)
-        else:
-            return r'{{%s}^\dag}' % str(self.name)
-
-    def _print_contents(self, printer, *args):
-        if self.is_annihilation:
-            return r'%s' % str(self.name)
-        else:
-            return r'Dagger(%s)' % str(self.name)
-
-    def _print_contents_pretty(self, printer, *args):
-        from sympy.printing.pretty.stringpict import prettyForm
-        pform = printer._print(self.args[0], *args)
-        if self.is_annihilation:
-            return pform
-        else:
-            return pform**prettyForm(u('\u2020'))
 
 
 def _expand_powers(factors):
@@ -513,204 +312,3 @@ def normal_order(expr, recursive_limit=10, _recursive_depth=0):
                                     _recursive_depth=_recursive_depth)
     else:
         return expr
-
-
-class BosonFockKet(Ket):
-    """Fock state ket for a bosonic mode.
-
-    Parameters
-    ==========
-
-    n : Number
-        The Fock state number.
-
-    """
-
-    def __new__(cls, n):
-        return Ket.__new__(cls, n)
-
-    @property
-    def n(self):
-        return self.label[0]
-
-    @classmethod
-    def dual_class(self):
-        return BosonFockBra
-
-    @classmethod
-    def _eval_hilbert_space(cls, label):
-        return FockSpace()
-
-    def _eval_innerproduct_BosonFockBra(self, bra, **hints):
-        return KroneckerDelta(self.n, bra.n)
-
-    def _apply_operator_BosonOperator(self, op, **options):
-        if op.is_annihilation:
-            return sqrt(self.n) * BosonFockKet(self.n - 1)
-        else:
-            return sqrt(self.n + 1) * BosonFockKet(self.n + 1)
-
-
-class BosonFockBra(Bra):
-    """Fock state bra for a bosonic mode.
-
-    Parameters
-    ==========
-
-    n : Number
-        The Fock state number.
-
-    """
-
-    def __new__(cls, n):
-        return Bra.__new__(cls, n)
-
-    @property
-    def n(self):
-        return self.label[0]
-
-    @classmethod
-    def dual_class(self):
-        return BosonFockKet
-
-    @classmethod
-    def _eval_hilbert_space(cls, label):
-        return FockSpace()
-
-
-class BosonCoherentKet(Ket):
-    """Coherent state ket for a bosonic mode.
-
-    Parameters
-    ==========
-
-    alpha : Number, Symbol
-        The complex amplitude of the coherent state.
-
-    """
-
-    def __new__(cls, alpha):
-        return Ket.__new__(cls, alpha)
-
-    @property
-    def alpha(self):
-        return self.label[0]
-
-    @classmethod
-    def dual_class(self):
-        return BosonCoherentBra
-
-    @classmethod
-    def _eval_hilbert_space(cls, label):
-        return HilbertSpace()
-
-    def _eval_innerproduct_BosonCoherentBra(self, bra, **hints):
-        if self.alpha == bra.alpha:
-            return Integer(1)
-        else:
-            return exp(-(abs(self.alpha)**2 + abs(bra.alpha)**2 -
-                         2 * conjugate(bra.alpha) * self.alpha)/2)
-
-    def _apply_operator_BosonOperator(self, op, **options):
-        if op.is_annihilation:
-            return self.alpha * self
-        else:
-            return None
-
-
-class BosonCoherentBra(Bra):
-    """Coherent state bra for a bosonic mode.
-
-    Parameters
-    ==========
-
-    alpha : Number, Symbol
-        The complex amplitude of the coherent state.
-
-    """
-
-    def __new__(cls, alpha):
-        return Bra.__new__(cls, alpha)
-
-    @property
-    def alpha(self):
-        return self.label[0]
-
-    @classmethod
-    def dual_class(self):
-        return BosonCoherentKet
-
-    def _apply_operator_BosonOperator(self, op, **options):
-        if not op.is_annihilation:
-            return self.alpha * self
-        else:
-            return None
-
-
-class FermionFockKet(Ket):
-    """Fock state ket for a fermionic mode.
-
-    Parameters
-    ==========
-
-    n : Number
-        The Fock state number.
-
-    """
-
-    def __new__(cls, n):
-        if n not in [0, 1]:
-            raise ValueError("n must be 0 or 1")
-        return Ket.__new__(cls, n)
-
-    @property
-    def n(self):
-        return self.label[0]
-
-    @classmethod
-    def dual_class(self):
-        return FermionFockBra
-
-    @classmethod
-    def _eval_hilbert_space(cls, label):
-        return HilbertSpace()
-
-    def _eval_innerproduct_FermionFockBra(self, bra, **hints):
-        return KroneckerDelta(self.n, bra.n)
-
-    def _apply_operator_FermionOperator(self, op, **options):
-        if op.is_annihilation:
-            if self.n == 1:
-                return FermionFockKet(0)
-            else:
-                return Integer(0)
-        else:
-            if self.n == 0:
-                return FermionFockKet(1)
-            else:
-                return Integer(0)
-
-
-class FermionFockBra(Bra):
-    """Fock state bra for a fermionic mode.
-
-    Parameters
-    ==========
-
-    n : Number
-        The Fock state number.
-
-    """
-
-    def __new__(cls, n):
-        if n not in [0, 1]:
-            raise ValueError("n must be 0 or 1")
-        return Bra.__new__(cls, n)
-
-    @property
-    def n(self):
-        return self.label[0]
-
-    @classmethod
-    def dual_class(self):
-        return FermionFockKet
